feat: Merge us_plot3d_xyz into us_plot3d
